LF-20668: Fixing missing twitter controls in IE11
https://livefyre.atlassian.net/browse/LF-20668
