+++ bower.json
@@ -1,6 +1,6 @@
 {
   "name": "streamhub-sdk",
-  "version": "2.19.1",
+  "version": "2.19.2",
   "main": [
     "./src/main.js"
   ],
