+++ src/content/views/content-view.js
@@ -11,6 +11,7 @@ var debug = require('debug');
 
 'use strict';
 
+var hasInnerHtmlBug = null;
 var log = debug('streamhub-sdk/content/views/content-view');
 
 /**
@@ -191,5 +192,43 @@ ContentView.prototype.destroy = function () {
     CompositeView.prototype.destroy.call(this);
     this.content = null;
 };
+
+/**
+ * Render the content inside of the ContentView's element.
+ * @returns {ContentView}
+ */
+ContentView.prototype.render = function () {
+    /**
+     * bengo:
+     * This next 3 lines makes me sad, but it is necessary to support IE9.
+     * View.prototype.render will set this.innerHTML to template().
+     * For some reason, this also causes the innerHTML of the buttons to
+     * be set to an empty string. e.g. Like Buttons have their like count
+     * cleared out. When ._renderButtons later re-appendChilds all the
+     * button.els, they are empty. So if we detach them here before
+     * this.innerHTML is set, they are not cleared.
+     * bit.ly/1no8mNk 
+     */
+    if (hasInnerHtmlBug = testHasInnerHtmlBug()) {
+        this._footerView._detachButtons();
+    }
+
+    CompositeView.prototype.render.call(this);
+    return this;
+};
+
+function testHasInnerHtmlBug() {
+    // only test once
+    if (hasInnerHtmlBug !== null) {
+        return hasInnerHtmlBug
+    }
+    var txt = 'hi';
+    var parent = document.createElement('div');
+    var child = document.createElement('span');
+    child.appendChild(document.createTextNode(txt));
+    parent.appendChild(child);
+    parent.innerHTML = '';
+    return child.innerHTML === '';
+}
 
 module.exports = ContentView;
